Move metrics tests to separate file
